moving keyset list into bottom sheet progress 2
