+++ android/src/main/java/io/parity/signer/screens/keysets/create/backupstepscreens/NewKeySetNetworksViewModel.kt
@@ -25,6 +25,7 @@ class NewKeySetNetworksViewModel : ViewModel() {
 		networkForKeys: Set<NetworkModel>,
 		onAfterCreate: (Boolean) -> Unit = {},
 	): Unit {
+		//todo dmitry check that we updating lastKnown in seedRepo
 		viewModelScope.launch {
 			val success = createKeySetUseCase.createKeySetWithNetworks(
 				seedName, seedPhrase,
